Fixed BitArray freeing, and improved to_s speed.
BitArray freeing is now done with ruby_xfree, so that the freed memory
is accounted for by the garbage collector. Not doing this before was an
oversite.

BitArray#to_s now works by creating a single C string and converting it
to a Ruby string in one go. This results in a significant speedup (from
more than 3 seconds to under 1 second for 10000 calls.) The formatting
is also changed to match that used by BitField.

Testing and benchmarking files are now under test/, along with a copy of
bitfield.rb for comparison.