Slicing should work as with Array now.
I also fixed the behavior of initialize when given a negative size.
Before, the size was interpreted as size_t, so giving it a -1 resulted
in creating a BitArray of the maximum size. Now it sets the length to 0.

I'm starting to think that I should replace all uses of size_t and
ptrdiff_t with long. It seems to work for Array.