Fixed the memory allocation problem.
The issue was the uint_array_size(bits) macro. When "bits" was 0, it
expanded to (-1 / UINT_BITS) + 1. Because UINT_BITS is an unsigned long,
-1 gets converted to a very large unsigned long, as per the standard. I
changed to macro to check for that, and also changed initialize_bitarray
to not bother allocating anything if it's given a size of 0.

Everything now uses Ruby's allocation functions, so garbage collection
should be triggered when necessary.

My theory is that the memory allocation wasn't an issue when using
glibc's malloc because of Linux's memory overcommit behavior; the memory
wasn't getting used, so it never actually got allocated. Apparently,
ruby_xmalloc and friends actually do touch the memory, so it all gets
allocated.